requirements_development.txt
-r requirements.txt
werkzeug
ipython
django-debug-toolbar
django-extensions
